# contracts/openapi.json
{"openapi":"3.0.1","info":{"title":"Jobseeker Copilot - Document Export API","description":"Exports generated document content into DOCX and PDF files and saves them\nto document-store-service. Every export operation requires the approved\nGateway service identity and its trusted document-owner context. Edited\nDOCX uploads become new document versions only after both authoritative\nDOCX and matching PDF files have been stored.\n","version":"2.1.0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"tags":[{"name":"Document Exports","description":"Endpoints for exporting generated documents to DOCX and PDF"}],"paths":{"/api/v1/document-exports/documents/{documentId}":{"post":{"tags":["Document Exports"],"summary":"Export generated document","description":"Creates DOCX and/or PDF exports and saves them to document-store-service","operationId":"exportDocument","parameters":[{"name":"X-Document-Owner","in":"header","description":"Owner context bound by the authenticated Gateway","required":true,"schema":{"type":"string"}},{"name":"documentId","in":"path","description":"UUID of the generated document","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentExportRequest"}}},"required":true},"responses":{"400":{"description":"Invalid export request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentExportResponse"}}}},"401":{"description":"Service authentication failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceIdentityError"}}}},"201":{"description":"Exports created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentExportResponse"}}}},"502":{"description":"Document store or export generation failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentExportResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/document-exports/documents/{documentId}/upload":{"post":{"tags":["Document Exports"],"summary":"Upload user-edited document file","description":"Imports an edited DOCX as a new inactive document version, stores the DOCX\nand a PDF rendered from the same imported text, then activates the version.\nReuse Idempotency-Key when retrying the same upload.\n","operationId":"uploadReplacement","parameters":[{"name":"X-Document-Owner","in":"header","description":"Owner context bound by the authenticated Gateway","required":true,"schema":{"type":"string"}},{"name":"Idempotency-Key","in":"header","description":"Stable retry key for this replacement operation","schema":{"maxLength":128,"type":"string"}},{"name":"documentId","in":"path","description":"UUID of the generated document","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"documentKind","in":"query","required":true,"schema":{"type":"string","enum":["CV","COVER_LETTER"]}},{"name":"uploadedFormat","in":"query","description":"Private beta replacements accept DOCX only. PDF remains in the\npublished client enum for wire compatibility and is rejected with 400.\n","required":true,"schema":{"type":"string","enum":["DOCX","PDF"]}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"502":{"description":"Document store failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentUploadResponse"}}}},"400":{"description":"Invalid upload request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentUploadResponse"}}}},"401":{"description":"Service authentication failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceIdentityError"}}}},"201":{"description":"Upload saved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentUploadResponse"}}}}},"security":[{"serviceToken":[]}]}}},"components":{"schemas":{"DocumentExportRequest":{"required":["formats"],"type":"object","properties":{"formats":{"type":"array","description":"Export formats to create","example":["DOCX","PDF"],"items":{"type":"string","description":"Export formats to create","example":"[\"DOCX\",\"PDF\"]","enum":["DOCX","PDF"]}}},"description":"Request to export a generated document into one or more formats"},"ServiceIdentityError":{"type":"object","properties":{"code":{"type":"string","example":"AUTHENTICATION_REQUIRED"},"message":{"type":"string","example":"Valid service authentication is required."}},"description":"Stable service-identity or owner-context rejection"},"DocumentExportItem":{"type":"object","properties":{"fileId":{"type":"string","format":"uuid"},"format":{"type":"string","enum":["DOCX","PDF"]},"fileName":{"type":"string"},"mimeType":{"type":"string"},"downloadUrl":{"type":"string"}},"description":"Single exported document file result"},"DocumentExportResponse":{"type":"object","properties":{"documentId":{"type":"string","format":"uuid"},"exports":{"type":"array","items":{"$ref":"#/components/schemas/DocumentExportItem"}}},"description":"Response containing exported files"},"DocumentUploadResponse":{"type":"object","properties":{"generatedDocumentId":{"type":"string","format":"uuid"},"uploadedFile":{"$ref":"#/components/schemas/DocumentExportItem"},"regeneratedFiles":{"type":"array","items":{"$ref":"#/components/schemas/DocumentExportItem"}},"latestFiles":{"$ref":"#/components/schemas/LatestDocumentFiles"},"message":{"type":"string"}},"description":"Response containing uploaded and latest document file metadata"},"LatestDocumentFiles":{"type":"object","properties":{"docx":{"$ref":"#/components/schemas/DocumentExportItem"},"pdf":{"$ref":"#/components/schemas/DocumentExportItem"}},"description":"Latest available DOCX and PDF files for a generated document"}},"securitySchemes":{"serviceToken":{"type":"apiKey","name":"X-Service-Token","in":"header"}}}}